Rename file figure to image
