+++ packages/gitbook-docs/src/markdoc/schema/image.ts
